pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["agent"]

[project]
name = "pulse"
version = "0.1.0"
description = "Weekly Product Review Pulse — automated app-store review analysis via MCP"
requires-python = ">=3.12"
dependencies = [
    "typer[all]>=0.12",
    "pydantic>=2.7",
    "pydantic-settings>=2.3",
    "pyyaml>=6.0",
    "structlog>=24.2",
    "jinja2>=3.1",
    "httpx>=0.27",
    "google-play-scraper>=1.2.7",
    "emoji>=2.15.0",
    "beautifulsoup4>=4.14.3",
    "lxml>=6.1.0",
    "sentence-transformers>=5.4.1",
    "umap-learn>=0.5.12",
    "hdbscan>=0.8.42",
    "keybert>=0.9.0",
    "numpy>=2.4.4",
    "openai>=2.32.0",
    "mcp>=1.27.0",
]

[project.scripts]
pulse = "agent.__main__:app"

[tool.uv]
package = true

[dependency-groups]
dev = [
    "pytest>=8.2",
    "ruff>=0.5",
    "mypy>=1.10",
    "types-pyyaml>=6.0",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.12"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"
markers = [
    "live: marks tests requiring live network or credentials (skip with -m 'not live')",
]
